Move LockState component NFC
